PyGC_Head:  Use "long double" instead of "double" as the worst-case alignment gimmick.  David Abrahams notes that the standard "long double" actually requires stricter alignment than "double" on some Tru64 box. On my box and yours <wink>, it's the same, so no harm done on most boxes.

+++ Include/objimpl.h
@@ -272,7 +272,7 @@ typedef union _gc_head {
 		union _gc_head *gc_prev;
 		int gc_refs;
 	} gc;
-	double dummy;  /* force worst-case alignment */
+	long double dummy;  /* force worst-case alignment */
 } PyGC_Head;
 
 extern PyGC_Head _PyGC_generation0;
